Move RepairInvalidToolCallsMiddleware into tool_errors
It belongs alongside the other tool-error middleware (ToolErrorMiddleware,
wrap_tool_errors) — all three turn a tool-call problem into a ToolMessage the
model can recover from, rather than living in a standalone one-class module.
Renames the test file to match.
